+++ src/lshade.cpp
@@ -130,7 +130,7 @@ void LSHADE::adaptParameters() {
     size_t ptemp;
     for (int i = 0; i < population.size(); ++i) {
         double fraction = 0.11;
-        ptemp= (round(fraction * population.size()));
+        ptemp= size_t(round(fraction * population.size()));
         if (ptemp<2 ) ptemp=2; 
         p[i] = ptemp;
     };
